+++ classes/class-lsx-videos-search.php
@@ -24,8 +24,8 @@ class LSX_Videos_Search {
 	private function __construct() {
 		add_filter( 'lsx_search_post_types', array( $this, 'enable_post_type' ), 100, 1 );
 		add_filter( 'lsx_search_post_types_plural', array( $this, 'post_type_plural' ), 100, 1 );
-		add_filter( 'lsx_search_categories', array( $this, 'enable_categories' ), 100, 1 );
-		add_filter( 'lsx_search_enabled', array( $this, 'enable_categories_search' ), 100, 1 );
+		// add_filter( 'lsx_search_categories', array( $this, 'enable_categories' ), 100, 1 );
+		// add_filter( 'lsx_search_enabled', array( $this, 'enable_categories_search' ), 100, 1 );
 		add_filter( 'lsx_search_prefix', array( $this, 'set_search_prefix' ), 100, 1 );
 	}
 
@@ -94,9 +94,9 @@ class LSX_Videos_Search {
 	 */
 	public function enable_categories_search( $enabled ) {
 		if ( is_tax( 'video-category' ) ) {
-			$enabled = true;
+			$enabled = false;
 		}
-		if ( is_post_type_archive( 'video' ) ) {
+		if ( isset( $this->options['display'][ $this->video_archive_enable_search ] ) && ( ! empty( $this->options ) ) && 'on' === $this->options['display'][ $this->video_archive_enable_search ] ) {
 			$enabled = true;
 		}
 		return $enabled;